Add mixer window with master and channel volume controls

+++ hotkeys.go
@@ -349,7 +349,7 @@ func wrapHotkeyInputs() {
 	metrics := goFace.Metrics()
 	linePx := math.Ceil(metrics.HAscent + metrics.HDescent + 2)
 	rowUnits := float32(linePx) / ui
-	padPx := float64(6 * ui)
+	padPx := float64(6) * float64(ui)
 
 	resize := func(it *eui.ItemData) {
 		if it == nil {
